# Use ugorji codec for unmarshalling responses from etcd server.