--- a/README.md
+++ b/README.md
@@ -29,6 +29,7 @@ func main() {
 
 ```bash
 go get github.com/coreos/go-etcd/etcd
+go get github.com/ugorji/go/codec
 ```
 
 ## Caveat