--- a/etcd/response.go
+++ b/etcd/response.go
@@ -1,10 +1,13 @@
 package etcd
 
+//go:generate codecgen -o response.generated.go response.go
+
 import (
-	"encoding/json"
 	"net/http"
 	"strconv"
 	"time"
+
+	"github.com/ugorji/go/codec"
 )
 
 const (
@@ -39,7 +42,7 @@ func (rr *RawResponse) Unmarshal() (*Response, error) {
 
 	resp := new(Response)
 
-	err := json.Unmarshal(rr.Body, resp)
+	err := codec.NewDecoderBytes(rr.Body, new(codec.JsonHandle)).Decode(resp)
 
 	if err != nil {
 		return nil, err