… etcd/response.generated.go