This change speeds up RawResponse.Unmarshal ~2x:

BEFORE:
```
$ go test -test.run=XXXXX -bench=.*
PASS
BenchmarkSmallResponseUnmarshal		   10000	    180181 ns/op
BenchmarkManySmallResponseUnmarshal	     100	  15115904 ns/op
BenchmarkMediumResponseUnmarshal	    1000	   2302976 ns/op
BenchmarkLargeResponseUnmarshal		      20	  94022971 ns/op
ok  	github.com/coreos/go-etcd/etcd	8.963s
```

AFTER:
```
$ go test -test.run=XXXXX -bench=.*
PASS
BenchmarkSmallResponseUnmarshal		   20000	     85873 ns/op
BenchmarkManySmallResponseUnmarshal	     200	   7069753 ns/op
BenchmarkMediumResponseUnmarshal	    1000	   1339490 ns/op
BenchmarkLargeResponseUnmarshal		      20	  60579560 ns/op
ok  	github.com/coreos/go-etcd/etcd	7.604s
```

I also updated install section in README.md file to make it clear that user have to ```go get``` ugorji codec.

Fixes #217 
Refs https://github.com/GoogleCloudPlatform/kubernetes/issues/4521

@davidopp @wojtek-t @xiang90 @timothysc