@xiang90 2015-06-03
@fgrzadkowski I am OK with this change. I am not sure how will this improve the overall perf of k8s.

A more interesting question is: do you also want to do this at server side? Why not let etcd server accept protobuf if you really care about perf?