@fgrzadkowski 2015-06-03
I think that moving towards protobuf makes sense, but it seems like a major change. This one is simple and from manual tests it improves latency metrics by 20-30%.