@timothysc 2015-06-03
@xiang90 crossref - https://github.com/GoogleCloudPlatform/kubernetes/issues/8132