@philips 2015-06-03
This is great. We should also do it for the still developing `github.com/coreos/etcd/client` package. Also, we should not vendor the codec package as go-etcd is a library.

Overall, LGTM.