@xiang90 2015-06-03
LGTM

/cc @yichengq you are more familiar with go-etcd codebase.